API documentation

VectorAlignments.lcs — Function

Find the longest common subsequence for a vector of sequences.

Examples

julia> lcs([1,3,5], [1,2,3])
2-element Vector{Any}:
 1
 3

lcs(v; norm, cf)

VectorAlignments.scs — Function

Find the shortest common supersequence for a vector of sequences.

Examples

julia> scs([1,3,5], [1,2,3])
4-element Vector{Any}:
 1
 2
 3
 5

scs(v; norm, cf)

VectorAlignments.align — Function

Align each of the vectors in the parameter list to the comprehensive SCS for all of the parameter vectors.

Examples

julia> align([1,3,5], [1,2,3])
2-element Vector{Any}:
 Any[1, nothing, 3, 5]
 Any[1, 2, 3, nothing]

align(v; norm, cf)

VectorAlignments.featurematrix — Function

Create a feature matrix for a list of vectors. In the resulting matrix, each column c represents the alignment of one vector with the SCS for the parameter vectors, each row r represents one feature, and each cell value r,c is the value for feature r in vector c.

Examples

julia> featurematrix([1,3,5], [4,5,6])
5×2 Matrix{Any}:
	1          nothing
	3          nothing
	nothing  4
	5         5
	nothing  6

featurematrix(v; norm, cf)